Use route groups to separate import layout from space layout
Move import pages into (import) route group with a minimal layout
(just EntityPageContentContainer), and space pages into (space) route
group with the full layout (cover, header, tabs, providers).

This eliminates SpaceLayoutChrome and ConditionalSpaceTabs — import
pages no longer inherit the space layout's data fetching or UI chrome.
