+++ apps/web/app/space/[id]/(import)/layout.tsx
@@ -1,0 +1,9 @@
+import { EntityPageContentContainer } from '~/partials/entity-page/entity-page-content-container';
+
+type LayoutProps = {
+  children: React.ReactNode;
+};
+
+export default function ImportLayout({ children }: LayoutProps) {
+  return <EntityPageContentContainer>{children}</EntityPageContentContainer>;
+}
