+++ apps/web/app/space/[id]/(space)/activity/page.tsx
