+++ apps/web/app/space/[id]/(space)/governance/page.tsx
@@ -21,7 +21,7 @@ import {
 import { GovernanceProposalsList } from '~/partials/governance/governance-proposals-list';
 import { GovernanceProposalsListInfiniteScroll } from '~/partials/governance/governance-proposals-list-infinite-scroll';
 
-import { cachedFetchSpace } from '../cached-fetch-space';
+import { cachedFetchSpace } from '../../cached-fetch-space';
 
 interface Props {
   params: Promise<{ id: string }>;
